CastSpellActiontTest, MuteActionTest, IgnoreActionTest,ClickModeTest
The tests above have been finished.For these four tests their
MaximunParams and MinimunParams should be tested.
Detailedly:
for CastSpellActionTest and IgnoreActionTest check if the process goes
within if or else conditions and test whether get the corresponding
values.
for MuteActionTest and ClickModeTest mock the interface to get the
latest eventLines to check if they are equal to the corresponding
Strings